+++ packages/pandas-gbq/docs/source/changelog.rst
@@ -10,6 +10,8 @@ Bug fixes
 ~~~~~~~~~
 
 - Encode floating point values with greater precision. (:issue:`326`)
+- Support ``INT64`` and other standard SQL aliases in
+  :func:`~pandas_gbq.to_gbq` ``table_schema`` argument. (:issue:`322`)
 
 
 .. _changelog-0.14.0: